type: optimization type
