+++ example/src/router/index.ts
@@ -5,8 +5,9 @@ import Layout from '../views/Layout.vue'
 import Table from '../views/Table.vue'
 import Form from '../views/Form.vue'
 import Crud from '../views/Crud.vue'
+import type { IRouteRecordRaw } from '/@src/index'
 
-const routes = [
+const routes: IRouteRecordRaw[] = [
   {
     path: '/',
     redirect: '/index',
