+++ src/types/public.ts
@@ -1,0 +1,67 @@
+import type { App, DefineComponent } from 'vue'
+import type { CrudMenu, FormMenu, IPagination } from './index'
+
+export type StringObject = Record<string, unknown>
+
+export type UnknownObject = Record<string | number, unknown>
+
+export type UnknownFunction = (...arg: unknown[]) => unknown
+
+export type DeepTypeof<T> = {
+  [Q in keyof T]: T[Q] extends UnknownObject ? DeepTypeof<T[Q]> | Q : Q
+}[keyof T]
+
+export type IDefineComponent<Props = UnknownObject> = DefineComponent<Props> & {
+  install: (app: App) => void
+}
+
+export type MenuOptions = CrudMenu & FormMenu & StringObject
+
+export interface InstallOptions extends StringObject {
+  /** Pagination Attributes */
+  pagination?: IPagination
+  /** Menu Attributes */
+  menu?: MenuOptions
+}
+
+export type IComponentSize = 'medium' | 'small' | 'mini' | undefined
+
+export type IScreenSize = 'xs' | 'sm' | 'md' | 'lg' | 'xl'
+
+export type IPlacementType =
+  | 'top'
+  | 'top-start'
+  | 'top-end'
+  | 'bottom'
+  | 'bottom-start'
+  | 'bottom-end'
+  | 'left'
+  | 'left-start'
+  | 'left-end'
+  | 'right'
+  | 'right-start'
+  | 'right-end'
+
+type IButtonType =
+  | 'primary'
+  | 'success'
+  | 'warning'
+  | 'danger'
+  | 'info'
+  | 'text'
+  | 'default'
+
+type IButtonNativeType = 'button' | 'submit' | 'reset'
+
+export interface IButtonProps {
+  type?: IButtonType
+  size?: IComponentSize
+  icon?: string
+  nativeType?: IButtonNativeType
+  loading?: boolean
+  disabled?: boolean
+  plain?: boolean
+  autofocus?: boolean
+  round?: boolean
+  circle?: boolean
+}
